restrição de usuarios

+++ routes/projeto_route.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, render_template, request, redirect, url_for
+from flask import Blueprint, render_template, request, redirect, url_for,session,flash
 from db import salvar_projeto, buscar_projeto, excluir_projeto, buscar_projeto_por_id, atualizar_projeto
 
 projeto_route = Blueprint('projeto_route', __name__)
@@ -20,17 +20,24 @@ def listar_projetos():
     projeto = buscar_projeto() 
     return render_template('projeto.html', projetos=projeto)  
 
+ 
 
-
-@projeto_route.route('/projeto/excluir/<int:id>', methods=['POST'])
+@projeto_route.route('/excluir/<int:id>', methods=['POST'])
 def excluir_projeto_route(id):
-    excluir_projeto(id)  
-    return redirect(url_for('projeto_route.listar_projetos'))  
+    if session.get('perfil') not in ['admin', 'gestor']:
+        return redirect(url_for('projeto_route.listar_projetos'))
+
+    excluir_projeto(id)
+    flash("Projeto excluído com sucesso.")
+    return redirect(url_for('projeto_route.listar_projetos'))
 
 
 
 @projeto_route.route('/projeto/editar/<int:id>', methods=['GET', 'POST'])
 def editar_projeto(id):
+    if session.get('perfil') not in ['admin', 'gestor']:
+        return redirect(url_for('projeto_route.listar_projetos'))
+    
     if request.method == 'POST':
         nome = request.form['nome']
         descricao = request.form['descricao']
